dash refatoração

+++ src/main/java/com/lafinance/dashboard/service/impl/CompraVendaServiceImpl.java
@@ -1,9 +1,14 @@
 package com.lafinance.dashboard.service.impl;
 
+import java.math.BigDecimal;
 import java.util.List;
+import java.util.Objects;
 
 import com.lafinance.dashboard.domain.model.Acao;
+import com.lafinance.dashboard.exception.BusinessException;
+import com.lafinance.dashboard.exception.NenhumRegistroEncontradoException;
 import com.lafinance.dashboard.service.AcaoService;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -12,44 +17,65 @@ import com.lafinance.dashboard.domain.model.CompraVenda;
 import com.lafinance.dashboard.repository.CompraVendaRepository;
 import com.lafinance.dashboard.service.CompraVendaService;
 
+import javax.inject.Inject;
+import javax.inject.Singleton;
+
 @Service
+@Singleton
 @Transactional
-public class CompraVendaServiceImpl implements CompraVendaService{
+@RequiredArgsConstructor(onConstructor = @__(@Inject))
+public class CompraVendaServiceImpl implements CompraVendaService {
 
-	@Autowired
-	private CompraVendaRepository repository;
+    public static final String ID_VENDA_NAO_INFORMADO = "Id venda não informado";
+    public static final String ID_NAO_INFORMADO = "ID não informado";
+    public static final String NENHUM_REGISTRO_ENCONTRADO_PELO_ID_INFORMADO = "Nenhum registro encontrado pelo id informado";
+    public static final String DADOS_DE_COMPRA_E_VENDA_NAO_INFORMADO = "Dados de compra e venda não informado";
 
-	@Autowired
-	private AcaoService acaoService;
+    private final CompraVendaRepository repository;
 
-	@Override
-	public void salvarRegistro(List<CompraVenda> compraVenda) {
-		repository.saveAll(compraVenda);
-		repository.flush();
-	}
+    @Autowired
+    private AcaoService acaoService;
 
-	@Override
-	public List<CompraVenda> consultarCompraVendaPeloIdVenda(Integer id) {
-		return repository.findByVendaId(id);
-	}
+    public void salvarRegistro(List<CompraVenda> compraVenda) throws Exception {
+        if (compraVenda.isEmpty())
+            throw new BusinessException(DADOS_DE_COMPRA_E_VENDA_NAO_INFORMADO);
 
-	@Override
-	public List<CompraVenda> consultarCompraVendaPeloIdCompra(Integer id) {;
-		return repository.findByCompraId(id);
-	}
+        repository.saveAll(compraVenda);
+        repository.flush();
+    }
 
-	@Override
-	public void excluirCompraVendaPeloIdVenda(Integer id) throws Exception {
-		try {
-			List<Acao> acaoList = this.acaoService.consultarAcoesPeloIdVenda(id);
-			this.acaoService.ativarAcoes(acaoList);
+    public void excluirCompraVendaPeloIdVenda(Integer id) throws Exception {
+        if (Objects.isNull(id))
+            throw new BusinessException(ID_NAO_INFORMADO);
 
-			List<CompraVenda> registros = this.repository.findByVendaId(id);
+        List<Acao> acaoList = this.acaoService.consultarAcoesPeloIdVenda(id);
 
-			this.repository.deleteInBatch(registros);
-		}catch (Exception e){
-			throw e;
-		}
-	}
+        if (acaoList.isEmpty())
+            throw new NenhumRegistroEncontradoException(NENHUM_REGISTRO_ENCONTRADO_PELO_ID_INFORMADO);
+
+        this.acaoService.ativarAcoes(acaoList);
+
+        List<CompraVenda> registros = this.repository.findByVendaId(id);
+
+        this.repository.deleteInBatch(registros);
+    }
+
+    public BigDecimal calcularLucroBrutoTotalPeloIdVenda(Integer idVenda) throws Exception {
+        if (Objects.isNull(idVenda))
+            throw new BusinessException(ID_VENDA_NAO_INFORMADO);
+
+        var lucroBrutoTotal = this.repository.calculoLucroBrutoPeloIdVenda(idVenda);
+
+        return lucroBrutoTotal;
+    }
+
+    public BigDecimal calcularLucroBrutoAno(Integer ano) {
+        return this.repository.calculoLucroAno(ano);
+    }
+
+    @Override
+    public List<CompraVenda> consultarCompraVendaPeloIdVenda(Integer id) {
+        return repository.findByVendaId(id);
+    }
 
 }
